Refactor: move Membership into a standalone mod
