+++ openraft/src/core/admin.rs
@@ -14,11 +14,11 @@ use crate::error::InitializeError;
 use crate::raft::AddLearnerResponse;
 use crate::raft::ClientWriteRequest;
 use crate::raft::ClientWriteResponse;
-use crate::raft::Membership;
 use crate::raft::RaftRespTx;
 use crate::AppData;
 use crate::AppDataResponse;
 use crate::LogId;
+use crate::Membership;
 use crate::NodeId;
 use crate::RaftError;
 use crate::RaftNetwork;
