+++ openraft/src/core/mod.rs
@@ -52,7 +52,6 @@ use crate::raft::ClientWriteRequest;
 use crate::raft::ClientWriteResponse;
 use crate::raft::Entry;
 use crate::raft::EntryPayload;
-use crate::raft::Membership;
 use crate::raft::RaftMsg;
 use crate::raft::RaftRespTx;
 use crate::replication::ReplicaEvent;
@@ -62,6 +61,7 @@ use crate::storage::InitialState;
 use crate::AppData;
 use crate::AppDataResponse;
 use crate::LogId;
+use crate::Membership;
 use crate::MessageSummary;
 use crate::NodeId;
 use crate::RaftNetwork;
