+++ openraft/src/membership/mod.rs
@@ -1,0 +1,7 @@
+#[allow(clippy::module_inception)]
+mod membership;
+
+#[cfg(test)]
+mod membership_test;
+
+pub use membership::Membership;
