+++ openraft/tests/snapshot/snapshot_overrides_membership.rs
@@ -6,9 +6,9 @@ use maplit::btreeset;
 use openraft::raft::AppendEntriesRequest;
 use openraft::raft::Entry;
 use openraft::raft::EntryPayload;
-use openraft::raft::Membership;
 use openraft::Config;
 use openraft::LogId;
+use openraft::Membership;
 use openraft::RaftNetwork;
 use openraft::RaftStorage;
 use openraft::SnapshotPolicy;
