+++ openraft/tests/snapshot/snapshot_uses_prev_snap_membership.rs
@@ -3,9 +3,9 @@ use std::time::Duration;
 
 use anyhow::Result;
 use maplit::btreeset;
-use openraft::raft::Membership;
 use openraft::Config;
 use openraft::LogId;
+use openraft::Membership;
 use openraft::MessageSummary;
 use openraft::RaftStorage;
 use openraft::SnapshotPolicy;
